实现批次删除功能

+++ src/main/java/com/final_back/impl/cultivation/DosingRecordImpl.java
@@ -43,6 +43,9 @@ public class DosingRecordImpl extends ServiceImpl<DosingRecordMapper, DosingReco
 
     @Override
     public int deleteDosingRecordByIdList(List<Long> idList) {
-        return dosingRecordMapper.deleteBatchIds(idList);
+        if (idList.size() > 0){
+            return dosingRecordMapper.deleteBatchIds(idList);
+        }
+        return 0;
     }
 }
